requirements.txt
streamlit
pillow
numpy
pandas
streamlit-image-coordinates
opencv-python